Club post, Joined Clubs, Buttons
Made it so that the data when creating a club post is saved and still displays when navigating away and back. The advertisement also has a join button.

Joined clubs will also display on the Joined Clubs page and also has a leave club button.

Back button added in Joined Clubs - Links back to community